deviding a programm into files
