+++ tetris2/tetris2/main.C
@@ -1,0 +1,19 @@
+#include <stdio.h>
+#include <conio.h>
+#include <string.h>
+#include <stdlib.h>
+#include <Windows.h>
+#include <time.h>
+#include "Tetris.h"
+
+
+
+void main(void)
+{
+	//srand((unsigned)time(NULL));
+	printf("%s%s", GAME_TITLE,
+		"Press any key to begin...\n");
+	_getch(); if (_kbhit()) _getch();
+
+	GameMenu();
+}
