+++ tetris2/tetris2/tetris2.vcxproj
@@ -146,7 +146,11 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="tetris2.c" />
+    <ClCompile Include="main.C" />
+    <ClCompile Include="TetrisFunctions.c" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Tetris.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
