+++ tetris2/tetris2/tetris2.vcxproj.filters
@@ -15,8 +15,16 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="tetris2.c">
+    <ClCompile Include="main.C">
       <Filter>Файлы исходного кода</Filter>
     </ClCompile>
+    <ClCompile Include="TetrisFunctions.c">
+      <Filter>Файлы исходного кода</Filter>
+    </ClCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Tetris.h">
+      <Filter>Заголовочные файлы</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
