+++ tetris2/tetris2/x64/Debug/tetris2.log
@@ -1,7 +1,5 @@
-<<<<<<< HEAD
-﻿  tetris2.vcxproj -> C:\Users\Admin\Documents\политех16-17\прога\GIT\Polytech.cs.2017.SpringProject\tetris2\x64\Debug\tetris2.exe
-  tetris2.vcxproj -> C:\Users\Admin\Documents\политех16-17\прога\GIT\Polytech.cs.2017.SpringProject\tetris2\x64\Debug\tetris2.pdb (Full PDB)
-=======
-﻿  tetris2.vcxproj -> C:\Users\User\Desktop\Polytech_SpringProject\Polytech.cs.2017.SpringProject\tetris2\x64\Debug\tetris2.exe
+﻿  main.C
+  TetrisFunctions.c
+  Создание кода...
+  tetris2.vcxproj -> C:\Users\User\Desktop\Polytech_SpringProject\Polytech.cs.2017.SpringProject\tetris2\x64\Debug\tetris2.exe
   tetris2.vcxproj -> C:\Users\User\Desktop\Polytech_SpringProject\Polytech.cs.2017.SpringProject\tetris2\x64\Debug\tetris2.pdb (Full PDB)
->>>>>>> 7cc535b6439aa7a84b909478f457d4dcc244cab8
